added LambdaDeployer developer mode support (to develop faster)

+++ tests/remotespecs/src/test/java/de/quantummaid/httpmaid/remotespecs/lambda/LambdaDeployer.java
@@ -33,10 +33,10 @@ import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;
 import lombok.RequiredArgsConstructor;
 import lombok.ToString;
+import software.amazon.awssdk.services.sts.StsClient;
+import software.amazon.awssdk.services.sts.model.GetCallerIdentityResponse;
 
 import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.IOException;
 import java.util.Map;
 import java.util.Optional;
 
@@ -48,24 +48,24 @@ import static de.quantummaid.httpmaid.remotespecs.lambda.aws.s3.S3Handler.upload
 import static de.quantummaid.httpmaid.remotespecs.lambda.aws.websocketapi.WebsocketApiHandler.loadWebsocketApiInformation;
 import static de.quantummaid.httpmaid.tests.givenwhenthen.deploy.Deployment.httpDeployment;
 import static java.util.Optional.ofNullable;
-import static java.util.UUID.randomUUID;
 
 /**
- * When stackIdentifier is not user-supplied, then
- * - the default stackIdentifier is: httpmaid-remotespecs-${AWS::AccountId}
- * - the default stack naming scheme is: ${stackIdentifier}-(lambda|bucket)
- * - The bucket is owned by ??? (whoever runs through the test setup first)
- * - Any other user needing access must have that access granted some other way
- *   (additional policy, etc...). Users who cannot do that simply define
- *   a user-supplied stackidentifier.
- * - cleanup policy
+ * When stackIdentifier is not user-supplied (developerMode == false), then
+ *  - the default stackIdentifier is: httpmaid-remotespecs-${AWS::AccountId}
+ *  - the default stack naming scheme is: ${stackIdentifier}-(lambda|bucket)
+ *  - The bucket is owned by ??? (whoever runs through the test setup first)
+ *  - Any other user needing access must have that access granted some other way
+ *    (additional policy, etc...). Users who cannot do that simply define
+ *    a user-supplied stackidentifier.
+ *  - cleanup policy
  *   - bucket stack remains
  *   - lambda stack is deleted
  *
- * When stackIdentifier is user-supplied, we're in the "per-user-infra" mode
- * - $stackIdentifier-lambda is created, owned by the user running the test
- * - $stackIdentifier-bucket is created, ditto
- * - cleanup policy
+ * When stackIdentifier is user-supplied (developerMode == true), we're in
+ * the "per-user-infra" mode.
+ *  - $stackIdentifier-lambda is created, owned by the user running the test
+ *  - $stackIdentifier-bucket is created, ditto
+ *  - cleanup policy
  *   - bucket stack remains
  *   - lambda stack remains
  *   (ie Both must be cleaned up / managed by the user)
@@ -74,7 +74,7 @@ import static java.util.UUID.randomUUID;
 @EqualsAndHashCode
 @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
 public final class LambdaDeployer implements RemoteSpecsDeployer {
-    private static final String PREFIX = "remotespecsX";
+    private static final String SHARED_STACK_PREFIX = "httpmaid-remotespecs";
     private static final String RELATIVE_PATH_TO_LAMBDA_JAR = "/tests/lambda/target/remotespecs.jar";
     private static final String BUCKET_NAME = "remotespecs";
     private static final String REALTIVE_PATH_TO_CLOUDFORMATION_TEMPLATE = "/tests/remotespecs";
@@ -82,13 +82,28 @@ public final class LambdaDeployer implements RemoteSpecsDeployer {
     private static final String HTTP_V2_PAYLOAD_API_NAME = " RemoteSpecs HTTP Api (Payload Version 2.0) Lambda Proxy";
     private static final String WEBSOCKET_API_NAME = " RemoteSpecs WebSockets Lambda Proxy";
 
-    private static final int PORT = 443;
     private final String stackIdentifier;
+    private final Boolean developerMode;
 
     public static LambdaDeployer lambdaDeployer() {
-        final String stackIdentifier = userProvidedStackIdentifier()
-                .orElse(PREFIX + randomUUID().toString());
-        return new LambdaDeployer(stackIdentifier);
+        final LambdaDeployer deployer = userProvidedStackIdentifier()
+                .map(LambdaDeployer::perUserLambdaDeployer)
+                .orElseGet(LambdaDeployer::sharedLambdaDeployer);
+        return deployer;
+    }
+
+    private static LambdaDeployer perUserLambdaDeployer(final String stackIdentifier) {
+        return new LambdaDeployer(stackIdentifier, true);
+    }
+
+    private static LambdaDeployer sharedLambdaDeployer() {
+        return new LambdaDeployer(sharedStackIdentifier(), false);
+    }
+
+    private static String sharedStackIdentifier() {
+        final GetCallerIdentityResponse callerIdentity = StsClient.create().getCallerIdentity();
+        final String accountId = callerIdentity.account();
+        return String.format("%s-%s", SHARED_STACK_PREFIX, accountId);
     }
 
     @Override
@@ -103,7 +118,7 @@ public final class LambdaDeployer implements RemoteSpecsDeployer {
         final String basePath = BaseDirectoryFinder.findProjectBaseDirectory();
         final String lambdaPath = basePath + RELATIVE_PATH_TO_LAMBDA_JAR;
         final File file = new File(lambdaPath);
-        uploadToS3Bucket(BUCKET_NAME, stackIdentifier, file);
+        uploadToS3Bucket(artifactBucketName, stackIdentifier, file);
 
         create("cf-lambda.yml", stackIdentifier + "-lambda",
                 Map.of("StackIdentifier", stackIdentifier,
@@ -133,9 +148,12 @@ public final class LambdaDeployer implements RemoteSpecsDeployer {
     }
 
     private void cleanUp() {
+        if (developerMode) {
+            return;
+        }
         deleteAllObjectsInBucket(BUCKET_NAME);
         try (final CloudFormationHandler cloudFormationHandler = connectToCloudFormation()) {
-            cloudFormationHandler.deleteStacksStartingWith(PREFIX);
+            cloudFormationHandler.deleteStacksStartingWith(SHARED_STACK_PREFIX);
         }
     }
 
@@ -149,7 +167,7 @@ public final class LambdaDeployer implements RemoteSpecsDeployer {
 
     private static Optional<String> userProvidedStackIdentifier() {
         final String stackIdentifier = System.getenv("REMOTESPECS_STACK_IDENTIFIER");
-        return ofNullable(stackIdentifier);
+        return ofNullable(stackIdentifier).map(s -> "dev-" + s);
     }
 
     private static void create(
